dagster-cloud-cli/pyproject.toml
[build-system]
requires = [
    "setuptools>=61.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "dagster-cloud-cli"
requires-python = ">=3.9,<3.13"
dynamic = [
    "version",
]
authors = [
    { name = "Elementl", email = "hello@elementl.com" },
]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "dagster==1.10.3",
    "packaging>=20.9",
    "questionary",
    "requests",
    "typer>=0.4.1",
    "PyYAML>=5.1",
    "github3.py",
]

[project.license]
text = "Apache-2.0"

[project.optional-dependencies]
tests = [
    "freezegun",
    "dagster[test]",
    "dagster-pipes",
]

[project.scripts]
dagster-cloud = "dagster_cloud_cli.entrypoint:app"
dagster-plus = "dagster_cloud_cli.entrypoint:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic.version]
attr = "dagster_cloud_cli.version.__version__"

[tool.setuptools.packages.find]
exclude = [
    "dagster_cloud.cli_tests*",
]
namespaces = false
